added function for user file upload to vector store
